✅ [test]: ApiProvider 로직 변경으로 인한 테스트 로직 변경
*불필요한 서버 테스트 코드 삭제
